Moving auth again

+++ src/controllers/comment.js
@@ -6,7 +6,7 @@ const mongoose = require('mongoose');
 exports.comment = (req, res, next) => {
 	const { username } = req;
 	const { postId, content } = req.body;
-	
+
 	if (!postId || !username || !content) {
 		return res.status(401).send({ error: true, requiredAttributes: {
 			postIdPresent: postId !== undefined ? true : false,
@@ -19,7 +19,9 @@ exports.comment = (req, res, next) => {
 			if (!returnedPost) {
 				res.status(401).send({ error: true, message: 'There is no post related to that ID!' })
 			}
-			checkAuth(res, returnedPost.user, username)
+			if (returnedPost.user !== username) {
+				return res.status(401).json({ error: true, message: 'You are not authorized to perform this action' })
+			}
 			const comment = new Comment({
 				content: content,
 				user: username,
@@ -73,7 +75,9 @@ exports.editComment = (req, res, next) => {
 			if (!comment) {
 				return res.status(404).json({ error: true, message: 'Could not find the comment'})
 			}
-			checkAuth(res, comment.user, username)
+			if (comment.user !== username) {
+				return res.status(401).json({ error: true, message: 'You are not authorized to perform this action' })
+			}
 			comment.edited = true;
 			comment.content = content;
 			comment.save().then(_ => {
@@ -100,7 +104,9 @@ exports.deleteComment = (req, res, next) => {
 			if (!comment) {
 				return res.status(404).json({ error: true, message: 'Could not find the comment'})
 			}
-			checkAuth(res, comment.user, username)
+			if (comment.user !== username) {
+				return res.status(401).json({ error: true, message: 'You are not authorized to perform this action' })
+			}
 			Comment.findByIdAndDelete(id).then(comment => {
 				Post.findOne(comment.parent).then(post => {
 					const index = post.comments.indexOf(comment.parent);
@@ -115,11 +121,5 @@ exports.deleteComment = (req, res, next) => {
 		})
 	} else {
 		res.status(404).send({ error: true, message: 'Invalid ID' })
-	}
-}
-
-function checkAuth(res, foundProperty, username) {
-	if (foundProperty !== username) {
-		res.status(401).json({ error: true, message: 'You are not authorized to perform this action' })
 	}
 }
